testing things

+++ shared/src/messages/mod.rs
@@ -3,5 +3,5 @@ pub mod macros;
 pub mod server;
 
 pub type SessionKey = [u8; 32];
-pub type SessionId = i64;
+pub type SessionId = i32;
 pub type UserId = String;
